Reduce build with 'make addon' on list of files, generated by 'git ls-files'

+++ Makefile
@@ -13,8 +13,16 @@ rpi:
 	echo "Cross compiling currently not work. Call this on Raspian only"
 	#cd native_sources; make rpi && make install
 
-addon:
-	cd .. ; zip --symlinks -r $(PROJECT).zip $(PROJECT) --exclude \*/.\* $(PROJECT)/native_sources/\*
+addon: clean
+	@# 1. Create list of addon files. Filter out uncommited data
+	@# and native_sources folder, but add root folder.
+	git ls-files | grep -v "native_sources\/" | grep -v ".gitignore" \
+		| sed -n -e "s/.*/$(PROJECT)\/\0/p" \
+		> /dev/shm/$(PROJECT).include
+	echo "$(PROJECT)/root/*" >> /dev/shm/$(PROJECT).include
+	@# 2. Create archive
+	cd .. ; zip --symlinks -r $(PROJECT).zip . \
+		-i@/dev/shm/$(PROJECT).include
 
 clean:
-	rm ../$(PROJECT).zip
+	test \! -f ../$(PROJECT).zip || rm ../$(PROJECT).zip
